Deleted xml schema and made new ones to see if that was causing the problems. Did lots of general debugging, and even some refactoring to make things work better. Still, it is broken, and I don't know what causes it.
No one else online has run into this issue, implying it's of my own error. But yet, the way in which it contradicts itself is killing me. See comments in TaskModelData To understand what I have tried and found, and as much as I have been able to narrow it down to